+++ server/src/server.ts
@@ -24,7 +24,10 @@ import {
 	MarkupContent,
 	DocumentSymbolParams,
 	SymbolKind,
-	DocumentSymbol
+	DocumentSymbol,
+	FoldingRangeParams,
+	FoldingRange,
+	FoldingRangeKind
 } from 'vscode-languageserver';
 
 import { TextDocument } from 'vscode-languageserver-textdocument';
@@ -372,11 +375,11 @@ async function computeDiagnostics(doc: TextDocument) {
 								const valrange = Range.create(Position.create(i,parsed[i][parsed[i].length-2].p),Position.create(i,parsed[i][parsed[i].length-2].p+parsed[i][parsed[i].length-2].c));
 								const valtext = doc.getText(valrange);
 								if (
-									(thistypedoc.name === "STRING" && (parsed[i][parsed[i].length-2].l !== 3 || parsed[i][parsed[i].length-2].s !== 10)) ||
-									(thistypedoc.name === "COSEXPRESSION" && (parsed[i][parsed[i].length-2].l !== 3 || parsed[i][parsed[i].length-2].s !== 10)) ||
-									(thistypedoc.name === "CLASSNAME" && (parsed[i][parsed[i].length-2].l !== 3 || parsed[i][parsed[i].length-2].s !== 10)) ||
-									(thistypedoc.name === "INTEGER" && (parsed[i][parsed[i].length-2].l !== 3 || parsed[i][parsed[i].length-2].s !== 9)) ||
-									(thistypedoc.name === "BOOLEAN" && (parsed[i][parsed[i].length-2].l !== 3 || parsed[i][parsed[i].length-2].s !== 9 || (valtext !== "1" && valtext !== "0")))
+									(thistypedoc.name === "STRING" && (parsed[i][parsed[i].length-2].l !== ld.cls_langindex || parsed[i][parsed[i].length-2].s !== ld.cls_str_attrindex)) ||
+									(thistypedoc.name === "COSEXPRESSION" && (parsed[i][parsed[i].length-2].l !== ld.cls_langindex || parsed[i][parsed[i].length-2].s !== ld.cls_str_attrindex)) ||
+									(thistypedoc.name === "CLASSNAME" && (parsed[i][parsed[i].length-2].l !== ld.cls_langindex || parsed[i][parsed[i].length-2].s !== ld.cls_str_attrindex)) ||
+									(thistypedoc.name === "INTEGER" && (parsed[i][parsed[i].length-2].l !== ld.cls_langindex || parsed[i][parsed[i].length-2].s !== ld.cls_num_attrindex)) ||
+									(thistypedoc.name === "BOOLEAN" && (parsed[i][parsed[i].length-2].l !== ld.cls_langindex || parsed[i][parsed[i].length-2].s !== ld.cls_num_attrindex || (valtext !== "1" && valtext !== "0")))
 								) {
 									let diagnostic: Diagnostic = {
 										severity: DiagnosticSeverity.Warning,
@@ -2012,7 +2015,8 @@ connection.onInitialize((params: InitializeParams) => {
 					edits: true
 				}
 			},
-			documentSymbolProvider: true
+			documentSymbolProvider: true,
+			foldingRangeProvider: true
 		}
 	};
 });
@@ -2483,7 +2487,7 @@ connection.onCompletion(
 			}
 		}
 		else if (
-			(prevline.slice(-1) === "." && prevline.slice(-2,-1) !== "," && prevline.slice(-2,-1) !== " " &&
+			(prevline.slice(-1) === "." && prevline.slice(-2,-1) !== "," && prevline.slice(-2,-1) !== " "  &&
 			thistoken !== 0 && (triggerlang === ld.cos_langindex || triggerlang === ld.cls_langindex)) ||
 			(prevline.slice(-2) === ".#" && triggerlang === ld.cos_langindex)
 		) {
@@ -4905,6 +4909,464 @@ connection.onDocumentSymbol(
 				else if (parsed[line][0].l === ld.cos_langindex && parsed[line][0].s === ld.cos_command_attrindex && firsttokentext.toLowerCase() === "routine") {
 					// This is the ROUTINE header line
 					routinename = doc.getText(Range.create(Position.create(line,parsed[line][1].p),Position.create(line,parsed[line][1].p+parsed[line][1].c)));
+				}
+			}
+		}
+
+		return result;
+	}
+);
+
+connection.onFoldingRanges(
+	(params: FoldingRangeParams) => {
+		const parsed = parsedDocuments.get(params.textDocument.uri);
+		if (parsed === undefined) {return null;}
+		const doc = documents.get(params.textDocument.uri);
+		if (doc === undefined) {return null;}
+		var result: FoldingRange[] = [];
+
+		var openranges: FoldingRange[] = [];
+		var inmultilinemacro: boolean = false;
+		var dotteddolevel: number = 0;
+		var injsonxdata: boolean = false;
+		for (let line = 0; line < parsed.length; line++) {
+			if (parsed[line].length === 0) {
+				if (openranges.length > 0 && openranges[openranges.length-1].kind === FoldingRangeKind.Comment) {
+					// Comment block ended, so close the range and append it to the result array if the range is more than one line
+					if (openranges[openranges.length-1].startLine < openranges[openranges.length-1].endLine) {
+						result.push(openranges[openranges.length-1]);
+					}
+					openranges.pop();
+				}
+				// Close any open dotted Do ranges
+				for (let idx = openranges.length-1; idx >= 0; idx--) {
+					if (openranges[idx].kind === "isc-dotteddo") {
+						openranges[idx].endLine = line-1;
+						result.push(openranges[idx]);
+						openranges.splice(idx,1);
+						dotteddolevel--;
+					}
+				}
+				continue;
+			}
+			if (
+				(parsed[line][0].l === ld.cls_langindex && parsed[line][0].s === ld.cls_desc_attrindex) ||
+				(parsed[line][0].l === ld.cos_langindex && parsed[line][0].s === ld.cos_dcom_attrindex)
+			) {
+				// This line is a UDL description or COS documentation comment
+
+				if (openranges.length === 0 || (openranges.length > 0 && openranges[openranges.length-1].kind !== FoldingRangeKind.Comment)) {
+					// Start a new range
+					openranges.push({
+						startLine: line,
+						endLine: line,
+						kind: FoldingRangeKind.Comment
+					});
+				}
+				else {
+					// Extend the existing range
+					openranges[openranges.length-1].endLine = line;
+				}
+			}
+			else {
+				// This line isn't a UDL description or COS documentation comment
+
+				if (openranges.length > 0 && openranges[openranges.length-1].kind === FoldingRangeKind.Comment) {
+					// Comment block ended, so close the range and append it to the result array if the range is more than one line
+					if (openranges[openranges.length-1].startLine < openranges[openranges.length-1].endLine) {
+						result.push(openranges[openranges.length-1]);
+					}
+					openranges.pop();
+				}
+				if (inmultilinemacro) {
+					// Check if the last token is a ##Continue
+					if (
+						parsed[line][parsed[line].length-1].l !== ld.cos_langindex || parsed[line][parsed[line].length-1].s !== ld.cos_ppf_attrindex ||
+						doc.getText(Range.create(
+							Position.create(line,parsed[line][parsed[line].length-1].p),
+							Position.create(line,parsed[line][parsed[line].length-1].p+parsed[line][parsed[line].length-1].c)
+						)).toLowerCase() !== "continue"
+					) {
+						// This is the end of a multi-line macro
+						var prevrange = openranges.length-1;
+						for (let rge = openranges.length-1; rge >= 0; rge--) {
+							if (openranges[rge].kind === "isc-mlmacro") {
+								prevrange = rge;
+								break;
+							}
+						}
+						if (openranges[prevrange].startLine < line) {
+							openranges[prevrange].endLine = line;
+							result.push(openranges[prevrange]);
+						}
+						openranges.splice(prevrange,1);
+						inmultilinemacro = false;
+					}
+				}
+				if (
+					parsed[line].length >= 2 &&
+					(parsed[line][0].l == ld.cos_langindex && parsed[line][0].s == ld.cos_ppc_attrindex) &&
+					(parsed[line][1].l == ld.cos_langindex && parsed[line][1].s == ld.cos_ppc_attrindex)
+				) {
+					// This line starts with a COS preprocessor command
+
+					const ppc = doc.getText(Range.create(Position.create(line,parsed[line][0].p),Position.create(line,parsed[line][1].p+parsed[line][1].c))).toLowerCase();
+					if (ppc === "#if" || ppc === "#ifdef" || ppc === "#ifndef" || ppc === "#ifundef") {
+						// These preprocessor commands always open a new range
+						openranges.push({
+							startLine: line,
+							endLine: line,
+							kind: "isc-ppc"
+						});
+					}
+					else if (ppc === "#elseif" || ppc === "#elif" || ppc === "#else") {
+						// These preprocessor commands always open a new range and close the previous one
+						var prevrange = openranges.length-1;
+						for (let rge = openranges.length-1; rge >= 0; rge--) {
+							if (openranges[rge].kind === "isc-ppc") {
+								prevrange = rge;
+								break;
+							}
+						}
+						if (openranges[prevrange].startLine < line-1) {
+							openranges[prevrange].endLine = line-1;
+							result.push(openranges[prevrange]);
+						}
+						openranges.splice(prevrange,1);
+						openranges.push({
+							startLine: line,
+							endLine: line,
+							kind: "isc-ppc"
+						});
+					}
+					else if (ppc === "#endif") {
+						// #EndIf always closes the previous range
+						var prevrange = openranges.length-1;
+						for (let rge = openranges.length-1; rge >= 0; rge--) {
+							if (openranges[rge].kind === "isc-ppc") {
+								prevrange = rge;
+								break;
+							}
+						}
+						if (prevrange >= 0 && openranges[prevrange].startLine < line-1) {
+							openranges[prevrange].endLine = line-1;
+							result.push(openranges[prevrange]);
+						}
+						openranges.splice(prevrange,1);
+					}
+					else if (ppc === "#define" || ppc === "#def1arg") {
+						// Check if the last token is a ##Continue
+						if (
+							parsed[line][parsed[line].length-1].l == ld.cos_langindex && parsed[line][parsed[line].length-1].s == ld.cos_ppf_attrindex &&
+							doc.getText(Range.create(
+								Position.create(line,parsed[line][parsed[line].length-1].p),
+								Position.create(line,parsed[line][parsed[line].length-1].p+parsed[line][parsed[line].length-1].c)
+							)).toLowerCase() === "continue"
+						) {
+							// This is the start of a multi-line macro definition
+							openranges.push({
+								startLine: line,
+								endLine: line,
+								kind: "isc-mlmacro"
+							});
+							inmultilinemacro = true;
+						}
+					}
+				}
+				else if (parsed[line][0].l == ld.xml_langindex) {
+					// This is a line of XML
+
+					// Loop through the line of XML tokens and look for tag delimiters
+					for (let xmltkn = 0; xmltkn < parsed[line].length; xmltkn++) {
+						if (parsed[line][xmltkn].l == ld.xml_langindex && parsed[line][xmltkn].s == ld.xml_tagdelim_attrindex) {
+							// This is a tag delimiter 
+							const tokentext = doc.getText(Range.create(
+								Position.create(line,parsed[line][xmltkn].p),
+								Position.create(line,parsed[line][xmltkn].p+parsed[line][xmltkn].c)
+							));
+							if (tokentext === "<") {
+								// Open a new XML range
+								openranges.push({
+									startLine: line,
+									endLine: line,
+									kind: "isc-xml"
+								});
+							}
+							else if (tokentext === "</" || tokentext === "/>") {
+								// Close the most recent XML range
+								var prevrange = openranges.length-1;
+								for (let rge = openranges.length-1; rge >= 0; rge--) {
+									if (openranges[rge].kind === "isc-xml") {
+										prevrange = rge;
+										break;
+									}
+								}
+								if (prevrange >= 0 && openranges[prevrange].startLine < line-1) {
+									openranges[prevrange].endLine = line-1;
+									result.push(openranges[prevrange]);
+								}
+								openranges.splice(prevrange,1);
+							}
+						}
+					}
+				}
+				else if (parsed[line].length > 1 && parsed[line][0].l == ld.cls_langindex && parsed[line][0].s == ld.cls_delim_attrindex) {
+					// This line starts with a UDL delimiter
+
+					const firsttwochars = doc.getText(Range.create(Position.create(line,0),Position.create(line,2)));
+					if (firsttwochars === "</") {
+						// Close the most recent Storage range
+						var prevrange = openranges.length-1;
+						for (let rge = openranges.length-1; rge >= 0; rge--) {
+							if (openranges[rge].kind === "isc-storage") {
+								prevrange = rge;
+								break;
+							}
+						}
+						if (prevrange >= 0 && openranges[prevrange].startLine < line-1) {
+							openranges[prevrange].endLine = line-1;
+							result.push(openranges[prevrange]);
+						}
+						openranges.splice(prevrange,1);
+					}
+					else if (firsttwochars.slice(0,1) === "<") {
+						// This is an XML open tag
+						// Only create a Storage range for it if it's not closed on this line
+						var closed = 0;
+						for (let stkn = 1; stkn < parsed[line].length; stkn++) {
+							if (
+								stkn !== parsed[line].length - 1 &&
+								parsed[line][stkn].l == ld.cls_langindex && parsed[line][stkn].s == ld.cls_delim_attrindex &&
+								parsed[line][stkn+1].l == ld.cls_langindex && parsed[line][stkn+1].s == ld.cls_delim_attrindex &&
+								doc.getText(Range.create(Position.create(line,parsed[line][stkn].p),Position.create(line,parsed[line][stkn+1].p+parsed[line][stkn+1].c))) === "</"
+							) {
+								closed = 1;
+								break;
+							}
+						}
+						if (!closed) {
+							openranges.push({
+								startLine: line,
+								endLine: line,
+								kind: "isc-storage"
+							});
+						}
+					}
+				}
+				else if (parsed[line].length === 1 && parsed[line][0].l === ld.cos_langindex && parsed[line][0].s === ld.cos_comment_attrindex) {
+					// This line contains a COS comment, so check if it's a region marker
+
+					var comment = doc.getText(Range.create(Position.create(line,parsed[line][0].p),Position.create(line,parsed[line][0].p+parsed[line][0].c))).toLowerCase();
+					if (comment.slice(0,2) === "//" || comment.slice(0,2) === "#;") {
+						comment = comment.slice(2).trim();
+						if (comment === "#region") {
+							// Create a new Region range
+							openranges.push({
+								startLine: line,
+								endLine: line,
+								kind: FoldingRangeKind.Region
+							});
+						}
+						else if (comment === "#endregion") {
+							// Close the most recent Region range
+							var prevrange = openranges.length-1;
+							for (let rge = openranges.length-1; rge >= 0; rge--) {
+								if (openranges[rge].kind === FoldingRangeKind.Region) {
+									prevrange = rge;
+									break;
+								}
+							}
+							if (prevrange >= 0) {
+								openranges[prevrange].endLine = line;
+								result.push(openranges[prevrange]);
+								openranges.splice(prevrange,1);
+							}		
+						}
+					}
+				}
+				else if (parsed[line][0].l == ld.cls_langindex && parsed[line][0].s == ld.cls_keyword_attrindex) {
+					// This line starts with a UDL keyword
+
+					const keytext = doc.getText(Range.create(Position.create(line,parsed[line][0].p),Position.create(line,parsed[line][0].p+parsed[line][0].c))).toLowerCase();
+					if (keytext === "xdata") {
+						// This line is that start of an XData block
+						for (let k = 3; k < parsed[line].length; k++) {
+							if (parsed[line][k].l == ld.cls_langindex && parsed[line][k].s == ld.cls_keyword_attrindex) {
+								// This is a UDL trailing keyword
+								const keytext = doc.getText(Range.create(
+									Position.create(line,parsed[line][k].p),
+									Position.create(line,parsed[line][k].p+parsed[line][k].c)
+								)).toLowerCase();
+								if (keytext === "mimetype") {
+									// An MimeType is specified
+									const mimetype = doc.getText(Range.create(
+										Position.create(line,parsed[line][k+2].p+1),
+										Position.create(line,parsed[line][k+2].p+parsed[line][k+2].c-1)
+									));
+									if (mimetype === "application/json") {
+										// This is the start of an XData block containing JSON
+										injsonxdata = true;
+									}
+									break;
+								}
+							}
+						}
+					}
+					else if (injsonxdata && keytext !== "xdata") {
+						// We've reached the next class member
+						injsonxdata = false;
+					}
+				}
+				else if (injsonxdata) {
+					// We're in a JSON XData block so look for opening/closing curly braces and brackets
+
+					for (let tkn = 0; tkn < parsed[line].length; tkn++) {
+						if (parsed[line][tkn].l === ld.javascript_langindex && parsed[line][tkn].s === ld.javascript_delim_attrindex) {
+							// This is a JSON bracket
+							const jb = doc.getText(Range.create(Position.create(line,parsed[line][tkn].p),Position.create(line,parsed[line][tkn].p+parsed[line][tkn].c)));
+							if (jb === "[" || jb === "{") {
+								// Create a new JSON range
+								openranges.push({
+									startLine: line,
+									endLine: line,
+									kind: "isc-json"
+								});
+							}
+							else if (jb === "]" || jb === "}") {
+								// Close the most recent JSON range
+								var prevrange = openranges.length-1;
+								for (let rge = openranges.length-1; rge >= 0; rge--) {
+									if (openranges[rge].kind === "isc-json") {
+										prevrange = rge;
+										break;
+									}
+								}
+								if (prevrange >= 0 && openranges[prevrange].startLine < line-1) {
+									openranges[prevrange].endLine = line-1;
+									result.push(openranges[prevrange]);
+								}
+								openranges.splice(prevrange,1);
+							}
+						}
+					}
+				}
+				else {
+					for (let tkn = 0; tkn < parsed[line].length; tkn++) {
+						if (parsed[line][tkn].l === ld.cos_langindex && parsed[line][tkn].s === ld.cos_jsonb_attrindex) {
+							// This is a JSON bracket
+							const jb = doc.getText(Range.create(Position.create(line,parsed[line][tkn].p),Position.create(line,parsed[line][tkn].p+parsed[line][tkn].c)));
+							if (jb === "[" || jb === "{") {
+								// Create a new JSON range
+								openranges.push({
+									startLine: line,
+									endLine: line,
+									kind: "isc-json"
+								});
+							}
+							else {
+								// Close the most recent JSON range
+								var prevrange = openranges.length-1;
+								for (let rge = openranges.length-1; rge >= 0; rge--) {
+									if (openranges[rge].kind === "isc-json") {
+										prevrange = rge;
+										break;
+									}
+								}
+								if (prevrange >= 0 && openranges[prevrange].startLine < line-1) {
+									openranges[prevrange].endLine = line-1;
+									result.push(openranges[prevrange]);
+								}
+								openranges.splice(prevrange,1);
+							}
+						}
+						if (tkn+1 > dotteddolevel && parsed[line][tkn].l === ld.cos_langindex && parsed[line][tkn].s === ld.cos_dots_attrindex) {
+							// This is the start of a dotted Do
+							dotteddolevel++;
+							openranges.push({
+								startLine: line-1,
+								endLine: line-1,
+								kind: "isc-dotteddo"
+							});
+						}
+						if (tkn === 0 && dotteddolevel > 0) {
+							// We're in a dotted Do, so check if the line begins with the correct number of dots
+
+							if (parsed[line].length >= dotteddolevel) {
+								for (let level = dotteddolevel-1; level >= 0; level--) {
+									if (parsed[line][level].l !== ld.cos_langindex || parsed[line][level].s !== ld.cos_dots_attrindex) {
+										// This dotted Do level is closed
+										var prevrange = openranges.length-1;
+										for (let rge = openranges.length-1; rge >= 0; rge--) {
+											if (openranges[rge].kind === "isc-dotteddo") {
+												prevrange = rge;
+												break;
+											}
+										}
+										openranges[prevrange].endLine = line-1;
+										result.push(openranges[prevrange]);
+										openranges.splice(prevrange,1);
+										dotteddolevel--;
+									}
+								}
+							}
+							else {
+								// At least one dotted Do level is closed
+								for (let level = dotteddolevel-1; level >= 0; level--) {
+									if (level > parsed[line].length-1) {
+										// Close all dotted Do levels that are greater than the length of this line
+										var prevrange = openranges.length-1;
+										for (let rge = openranges.length-1; rge >= 0; rge--) {
+											if (openranges[rge].kind === "isc-dotteddo") {
+												prevrange = rge;
+												break;
+											}
+										}
+										openranges[prevrange].endLine = line-1;
+										result.push(openranges[prevrange]);
+										openranges.splice(prevrange,1);
+										dotteddolevel--;
+									}
+									else if (parsed[line][level].l !== ld.cos_langindex || parsed[line][level].s !== ld.cos_dots_attrindex) {
+										// This dotted Do level is closed
+										var prevrange = openranges.length-1;
+										for (let rge = openranges.length-1; rge >= 0; rge--) {
+											if (openranges[rge].kind === "isc-dotteddo") {
+												prevrange = rge;
+												break;
+											}
+										}
+										openranges[prevrange].endLine = line-1;
+										result.push(openranges[prevrange]);
+										openranges.splice(prevrange,1);
+										dotteddolevel--;
+									}
+								}
+							}
+						}
+						if (parsed[line][tkn].l === ld.cos_langindex && parsed[line][tkn].s === ld.cos_embo_attrindex) {
+							// This is an embedded code block open token
+							openranges.push({
+								startLine: line,
+								endLine: line,
+								kind: "isc-embedded"
+							});
+						}
+						if (parsed[line][tkn].l === ld.cos_langindex && parsed[line][tkn].s === ld.cos_embc_attrindex) {
+							// This is an embedded code block close token
+							var prevrange = openranges.length-1;
+							for (let rge = openranges.length-1; rge >= 0; rge--) {
+								if (openranges[rge].kind === "isc-embedded") {
+									prevrange = rge;
+									break;
+								}
+							}
+							if (prevrange >= 0 && openranges[prevrange].startLine < line-1) {
+								openranges[prevrange].endLine = line-1;
+								result.push(openranges[prevrange]);
+							}
+							openranges.splice(prevrange,1);
+						}
+					}
 				}
 			}
 		}