+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [1.0.5] - 2020-XX-XX
 - Fix issue [#52](https://github.com/intersystems/language-server/issues/52): Request textDocument/documentSymbol failed with message: Cannot read property 'p' of undefined
+- Fix issue [#53](https://github.com/intersystems/language-server/issues/53): Become a FoldingRangeProvider
 - Fix issue [#54](https://github.com/intersystems/language-server/issues/54): Completion request fails when "." is typed as first non-whitespace character on a line
 - Fix issue [#55](https://github.com/intersystems/language-server/issues/55): Hover fails on Embedded SQL identifier that is the first word on the line